Move everything to the sidebar
