+++ src/Design/Nav/CategoryTag/CategoryButton.scss
@@ -1,4 +1,4 @@
-@import "../../../global";
+@import "../../global";
 
 .category-button {
   color: $a-color;
