+++ src/Design/Posts/PostList/PostList.scss
@@ -1,11 +1,12 @@
 @import "../../global";
 
 .container {
+  border: solid 1px;
   margin: 0 auto;
-  padding-top: 80px;
-  max-width: 1100px;
+  width: 1450px;
   .post-list-container {
-    float: left;
+    border: solid 1px;
+    margin: 80px 0 0 300px;
     min-height: 400px;
   }
 }
