--- src/Design/Posts/PostList/RightContainer/RightContainer.scss
@@ -1,116 +1,0 @@
-@import "../../../global";
-
-.right-container {
-  position: sticky;
-  top: 70px;
-  float: right;
-  width: 280px;
-}
-.search-box {
-  margin-top: 40px;
-  .search-box-title {
-    position: relative;
-    border: solid 0.5px $border-color2;
-    font-size: 23px;
-  }
-  .title {
-    position: absolute;
-    padding: 0 10px 0 10px;
-    background-color: #fff;
-    top: -15px;
-    right: 75px;
-  }
-  .search-bar {
-    border: none;
-    padding-left: 15px;
-    margin: 25px 5px 15px 5px;
-    outline: none !important;
-    border: solid 1px$border-color;
-    border-radius: 3px;
-    width: 270px;
-    height: 35px;
-    font-size: 17px;
-    transition: 1s;
-  }
-  .search-bar:focus {
-    border: solid 2px $border-color;
-  }
-}
-
-.tag-list-box {
-  margin-top: 40px;
-  .tag-list-title {
-    position: relative;
-    border: solid 0.5px $border-color2;
-    font-size: 23px;
-  }
-  .title {
-    position: absolute;
-    padding: 0 10px 0 10px;
-    background-color: #fff;
-    top: -15px;
-    right: 105px;
-  }
-  .tag-list {
-    margin-top: 25px;
-  }
-  .tag {
-    color: $p-color;
-    margin: 3px;
-    outline: none !important;
-    border: solid 1px $border-color;
-    border-radius: 3px;
-  }
-  .tag-active {
-    @extend .tag;
-    background-color: $p-color;
-    font-weight: bold;
-    color: #fff;
-  }
-  .tag:hover {
-    color: $h-color;
-  }
-  .tag-active:hover {
-    color: #fff;
-  }
-}
-
-.featured-post-box {
-  margin-top: 40px;
-  border-radius: 3px;
-  .featured-posts-title {
-    position: relative;
-    font-size: 23px;
-    border: solid 0.5px $border-color2;
-  }
-  .title {
-    position: absolute;
-    padding: 0 10px 0 10px;
-    background-color: #fff;
-    top: -15px;
-    right: 55px;
-  }
-  .featured-post {
-    margin: 30px 0 0 10px;
-    .post-title {
-      text-decoration: none;
-      font-size: 17px;
-      color: $theme-color;
-    }
-    .post-title:hover {
-      color: $a-color;
-    }
-  }
-  .featured-post-description {
-    color: $p-color;
-    font-size: 12px;
-  }
-}
-@media (max-width: 1200px) {
-  .right-container {
-    display: none;
-  }
-  .author-box {
-    display: none;
-  }
-}
